[refactor] : CommonFrame 추상메소드 추가
